requirements.txt
alembic==1.12.0
blinker==1.6.3
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
dominate==2.8.0
email-validator==2.0.0.post2
Flask==3.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login @ git+https://github.com/maxcountryman/flask-login.git@2204b4eee7b215977ba5a1bf85e2061f7fa65e20
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.0
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.2
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
visitor==0.1.3
Werkzeug==3.0.0
WTForms==3.1.0
